Testing: Extract test_header_units_and_modules.hpp.
